msm-4.14: Fix GCC warnings

+++ fs/f2fs/super.c
@@ -4586,8 +4586,8 @@ static int __init init_f2fs_fs(void)
 	int err;
 
 	if (PAGE_SIZE != F2FS_BLKSIZE) {
-		printk("F2FS not supported on PAGE_SIZE(%lu) != BLOCK_SIZE(%d)\n",
-				PAGE_SIZE, F2FS_BLKSIZE);
+		printk("F2FS not supported on PAGE_SIZE(%lu) != BLOCK_SIZE(%lu)\n",
+				PAGE_SIZE, (unsigned long)F2FS_BLKSIZE);
 		return -EINVAL;
 	}
